requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
boto3
python-dotenv
requests
